Use Global Gem Install
Durring the setup the "sass" gem is installed. This works fine on Linux operating systems like Ubuntu. However, on some other distrobutions such as Arch Linux "gem install" defaults to the a local folder in the users home directory. When installed here the sass binary is not on the path and coco-brunch cannot find it (and thus crashes with an error). By adding the "--no-user-install" flag we are telling gem to install it globally no matter what the system default is thus fixing this problem.