Fix Python Download
bootstrap.sh was using the "open" command to open a link. While this may work on a Mac, the command does not exist on Linux. Instead we can now use a switch to use the correct command for the users OS.

Ohh and python.org/getit is a 404 -> Changed it to python.org/download