Box2D in own script, rename Vector
Renaming Vector to Box2DVector to avoid naming collisions when
including Box2D on the main thread.  There may be a way to change how
scripts are built and included to avoid this rename, but it wasn’t
obvious.

Box2D declares a Vector in the global namespace, which conflicts with
the CodeCombat Vector.