Improved `updateLevelRequiredItems` logic regarding rings, fixes #2740
This fixes the main issue of #2740, allowing the player to play the level if the required ring is equipped on either ring slot.

There's still a small problem: if the player already had a ring equipped in the `right-ring` slot, the Ring of flowers will not display as required due to this [line](https://github.com/codecombat/codecombat/blob/0df6da7efa3df9ff6cd0e65133ff31923a9a0ee8/app/views/play/menu/InventoryModal.coffee#L436). This seems to be a different issue though, probably related to the `unequipLevelRestrictedItems` logic.