fixed some errors and added a few new tranlations
As I played through some of the levels to see if I would suggest this
game to my students, I encountered multiple translations that were
either phrased very weirdly, grammatically incorrect or completely
nonsensical.

I fixed those and also translated a few new lines along the way.

The german word for buy ("kaufen") is shorter than the word for unlock
("freschalten") and should fit in the button. I suggest changing that,
but it feels like avoiding the word buy was a design decision, so I left
it as it is.