Add saving/loading systems and components in branches
Make an interface for taking any group of locally modified systems and components
and saving them temporarily to the db. This is so the content team can shelve and/or
backup their work on the game engine, since it's a bunch of code that's in the db
where git can't get at it.