update arenas.md - add team names to colors
Just added "(the Humans" and "the Ogres" to make it clear what team "Red or Blue" means:

Students must choose a team to join: Red (the Humans) or Blue (the Ogres). It is important to divide the class as most students will choose red, if undirected. It doesn’t matter if the sides are even, but you need players on both sides in order to let your students play against one another. Ways to divide the class: