Shore up loading level sessions
When the CampaignView preloads a session, then that session
doesn't get used by the PlayLevelView. Tweaked it to make sure
the preloaded session is used, which might help:

* stop the choose hero modal from showing up
* changes to hero equipment sticking

And also possibly help the times when the play level view does
not load. Added logs for that also to figure out when that happens.