Update pt-BR.coffee [subscribe]
I only updated the 'subscribe' section here.
I improved some translations but the main change was on the mixing translations for the word 'subscribe' and its contractions:
I believe the best translation for "subscription" would be "assinatura" instead of "inscrição". The former is used in magazine and cable TV subscriptions context, for example. I found that both words were being used and I decided to change everything to "assinatura".