Update nl.coffee (#4891)
* Update nl.coffee

I added and changed some translations. I will work on this further in the coming days.

* Update nl.coffee (translations)

Hopefully the mistakes are now corrected. Let me know what else I can change if they are not corrected... thank you in advance!
I translated more, will work on this more in the future.