Remove unregisterModule from VueComponentView
When navigating to another vue component page, the destroy method for the previous page is executed after the new page is instantiated. So unregistering module in the destroy method unregisters it for the new page as well.
Therefore, each vue component should be responsible for registering and unregistering its own module having a unique name, instead of a generic 'page' module.