Added a PWA manifest file and ios touch icon
Currently there's no icon when you add CodeCombat to the home screen
on iOS (just a screenshot of the page). Also, the name of the icon on
the home screen is just the title of the page ("Play CodeCombat
Levels...").

I've added a high-res favicon for iOS, and I've also added a PWA
manifest file so that the icon title defaults to "CodeCombat". This
should also work on Android (although I haven't been able to test it
myself). On iOS I've also configured the app to hire the address bar,
when you start it from a home-screen icon, which gives a bit more
vertical space on the page and should make it feel more like a real
app.

There are other css issues with CodeCombat on iOS (which I or someone
else may fix some day) but I thought it would be easy to fix these
right away.