Update chalkboard editor to support rich editing of text colors (#556)
This adds a rich edit modal dialog for cinematics chalkboard content, which uses the third party Quill text editor.  The existing inline markdown ace editor is still in place, and should continue to work.

The new rich text content is now left-aligned instead of center-aligned.  The old markdown content format is still available, and still defaults to center-aligned.  Existing cinematics should continue working, although the cinematics UX was updated for this work.

The coupling between cinematics editor and treema-ext is not that great.

Image support for rich text modal editor works the same, and is a straight copy.