Fix large screen font size bug in Chalkboard (#616)
Adds a small CSS tweak to clamp chalkboard text to 24px. Zooming in and out may still cause issues however this improves the default 100% zoom experience on large screens.

I tested manually and will check again when it is on staging.

Refactored to CSS because scss preprocessor doesn't handle px and vmax!
Prior to this commit text can leak out of the chalkboard on a large screen size.