Update mn.coffee
Adding the translations into the mn.coffee file instead of the mon.coffee from now on. I think I need to do it all over again because the code's changed considerably since my first commit in 2015, but I'm improving on Google Translate, so should be easy to batch them by the hundreds on my weekends.