Remove unused short property names from AnalyticsLogEvent schema
In 2015, we were going to switch to u, e, p from user, event,
properties, and we started tracking some events both ways for a while
in preparation for switching, but then we never finished
migrating. Then, in 2017, we refactored the server and it stopped
storing the shortened event names anyway. We don't have events
currently from before 2018, so don't think we need this confusion in
the schema any more.

To shorten? No, don't want to have to deal with migrating data, and
MongoDB has long since gotten its compression on field names pretty
good, obviating the need for manual optimization to short field names.