Add experimental screen reader support option (#815)
This visualizes the map as an ASCII grid with a 5-meter resolution,
adds some screen-reader-only text updates, and tweaks how error
messages are displayed to include the message as well as the hint in
some cases.

It would be difficult to play the game with a screen reader in this
state, as I've only tested with visualizing the first few levels, and
there are many non-accessible UI elements that are necessary. But, as
a proof of concept, I think it shows that it can be technically
possible to understand the spatial layout of each level.