Update apcsp-search-sort.md
Addition
Standard: AAP-2.P.1, AAP-2.P.2, AAP-2.P.3.
Add to section: Search Algorithms.
Binary search: Open to the middle of the book. Determine which half of the book the word is in. Take that half, look at its center point, and repeat until you find the word. Note: although this search is often more efficient than sequential/linear searches, the data needs to be sorted first. 


Addition
Standard: AAP-4.A.5.
Add to section: Sorting Algorithms.

There are different ways to assess how efficient an algorithm is at sorting. For example, you can informally measure an algorithm’s efficiency by determining the number of times a statement or group of statements is executed. 

Most ways to sort lists work in a reasonable amount of time, even if they’re slower than others (with the exception of things like Bogo Sort). 

Addition
Standard: AAP-4.A.9.
Add to section: Sorting Algorithms.
These methods tend to involve “heuristic functions”, which approximate some part of the algorithm. A heuristic is an approach to a problem that produces a solution that is not guaranteed to be optimal but may be used when techniques that are guaranteed to always find an optimal solution are impractical.