Update apcsp-crowdsourcing.md
Addition
Standard: IOC-1.B.6
Add to section: How do you sort reviews or votes?

 But you also want to give “new” items a chance. Reddit addresses this by having separate tabs for popular and new items, so users can decide what they’d rather sort for.

It is important to note that while the rapid sharing of technology has been widely helpful in shared collaboration and the interchange of information, it still has some disadvantages beyond the intended purpose or control of the programmer. For instance, in the case of AirBnB, a sabotaging review might cause the property 
value to severely decline.