Update cs3.md
Addition
Standard: AAP-2.D.1.
Add to section: Instructive Activity: String Chuck.

Strings are pieces of text inside quotes, and a substring is part of an existing string.  The students have been using strings ever since Course 1. For example, in the buildXY() function, students use the string "fence", to build a fence as in hero.buildXY("fence", 34, 30). In the attack() function, the can choose to attack a chest by passing the string, "Chest" as a parameter, with hero.attack("Chest").

Addition
Standard: AAP-2.C.4.
Add to section: module 12, computer arithmetic, summary.

Computers can be used to add, subtract, multiply, and divide numbers. Additionally, they can be used to perform operations on variables representing numbers, and the results of functions that return numbers. It is important to note that, like calculators, computers apply the order of mathematical operations when evaluating mathematical expressions. 

Here’s the order that you should follow when configuring any mathematical expression:  Parentheses, Exponents, Multiplication and Division (from left to right), Addition and Subtraction (from left to right).