Fix Oz PlayLevelView keyboard navigation
This fixes several keyboard accessibility issues on the PlayLevelView,
its code editor, and some of its modals.

Some of this is using the new focus-trap package, which is something
we should be generally using on all modals so that keyboard tab
navigation doesn't leak out of the modal to all the content behind the
modal. This only applies it to the ModalTransition (victory modal) for
now. A general solution for all modals would be a good next step.

This also adds a top-level screen reader mode enable button that is
visible to screen readers, so you don't have to go looking through the
game options to find it.