Add level editor tabs for key thangs like Hero and Referee
When the level editor starts, if there is already a Hero Placeholder,
Hero Placeholder 1, Referee, Level Manager JS, etc., then this adds a
new convenience tab to edit that Thang, so it's easier to directly
jump to its configuration and to keep persistent code editor windows
open for use of external tooling.

This also adds a convenience function for automatically opening
Programmable Components to the code and merges a couple minor
coco/ozar differences (like not loading archived ThangTypes).