Fix cinematic music in Ozaria
We usually have an .mp3 with the same filename, but in many cinematics
we didn't specify it. Safari cannot play .ogg, so we need to specify
the .mp3 file. This automatically adds a parallel .mp3 entry if it was
missing, which should load music most of the time.

Also set cinematic music looping default to true (it was specified
manually everywhere, and true in all but two cinematics).