Fix CoCo teacher dashboard shared class visibility bugs
Previously, "Shared Classes" showed up as an empty section even if
there were no shared classes. Also, viewing as an admin showed shared
classes as if they were owned by the teacher in question.

Now, the "Shared Classes" section only shows if there are shared
classes, and the section grouping works when viewing as an admin.