Create a unified Footer component
This lets us clean up both .coco and .ozar variations of
base-flat.pug, base.pug, and style-flat.sass, plus removed a bunch of
other unused or partially unused lines. This is modeled after the top
Navigation Component. I've also removed obsolete code for the top nav.

Now CodeCombat and Ozaria will share the same footer, which largely
expands the Ozaria footer and puts it in the CodeCombat style.

Note: China footer is an example for how it could be done, but
koudashijie.com isn't running this footer code, so whatever
customizations it has should be merged into this with careful
conditionals so that we don't have to maintain two versions again.