Change top nav from "Educators" to "Dashboard" for teachers
When signed in as a teacher, it's clearer to have it be a click to
your teacher dashboard.

Also have the "Professional Development" submenu item not show up on
Ozaria if you're in the teacher dashboard, since there's already a PD
tab in your teacher dashboard. That item also links to the teacher
dashboard embedded version of the PD page instead of the root-level
marketing version of it.

On Ozaria, the "Educators" top-level nav link goes to / instead of #,
because you might not be on the homepage.