Give better error info and recovery options when levels can't load
Added the error message that caused the level to not load, instead of
falsely assigning it to an infinite loop when we know it's not.

There were some issues where Reset Level and Comment Out My Code
options worked on the second try; added a delay so that they work the
first try.

The Try Again option now doubles the infinite loop timeouts each time,
for those very long-simulating matches you really want to view.