Get rid of obsolete CHAPTER_TWO_PLAYTEST course ID references
In my testing, new CourseInstances for this stopped being created in
April, only admin teachers (of which we have two) saw them, and the
only change is that now they are last in the course ordering.

There are still a couple thousand CourseInstances pointing to the
'5eb34fc8dc0fd35e8eae66b0' course, which perhaps we could remove if it
turned out to be ok that some students might still have
`user.courseInstances` pointing at them.

The main effect is to get rid of the CH2P entry in the customized
license assignment modal for Ozaria.