Remove Ned Fulmer
This PR removes the ability to obtain Ned Fulmer. We still need to figure out how we should change/remove the custom pages for Ned. Also, upon pushing this PR, we need to set Ned Fulmer's releasePhase property to "beta" so it doesn't show up on the hero selection page.