Upgrade SpellPaletteView, make PlayLevelView resizing smarter, etc
Bring back the old below-editor code documentation area when there
aren't many APIs, modernizing it in the process. This gives us more
space for the Surface game area. When in this mode, method docs do
update with their interpolated property values, since they are
generated just-in-time. Defaults to bottom positioning when there are
fewer than 16 APIs, or you can use ?apis=bot or ?apis=mid to force the
alternative layout.

Fix up a lot of dynamic sizing code and styling to make better use of
the available screen space, including letting code editor flex to as
small as is needed for 80 characters or as wide as half the screen to
help either give the Surface more space or to keep it from overflowing
the bottom of the window.

Remove various dead code areas and fix up minor layout issues within
PlayLevelView for CodeCombat. Post-comment insertion point yellow
gutter arrows are back and visible, the Classrom JS|PY button is gone,
GameDevTrackView isn't erroneously pulling in Ozaria styling, error
message positioning is better, the LevelDialogueView now overlays the
playback bar and HUD rather than the Surface, the goals area doesn't
get so big so eagerly, and many elements are now positioned more
reliably near the edes of the Surface or code area.