Create a unified Footer component (#6806)
* Create a unified Footer component

This lets us clean up both .coco and .ozar variations of
base-flat.pug, base.pug, and style-flat.sass, plus removed a bunch of
other unused or partially unused lines. This is modeled after the top
Navigation Component. I've also removed obsolete code for the top nav.

Now CodeCombat and Ozaria will share the same footer, which largely
expands the Ozaria footer and puts it in the CodeCombat style.

This also merges in the footer modifications we have been using in China.

Co-authored-by: smallst_wsl <yuqiang_tong@163.com>