Allow teachers to enable items, gems, and exp in their classrooms (#6805)
* Allow teachers to enable items, gems, and exp in their classrooms

This fixes up the older experimental classroom items feature flag
and hooks it up to a teacher-controlled classroom setting. Now,
students playing for a classroom with this setting enabled will be
shown gem and exp rewards, be prompted to buy equipment in the item
shop, and gain stats and abilities according to their equipment
instead of just the default hero configuration in the level.

Later, if most teachers like this, we can default this to on by
changing the classroom schema default value for `classroomItems`.

It also introduces clampedProperties to levels, beyond just max/min
health, to let us further tune levels that are sensitive to
attackDamage, maxSpeed, and other properties.

This refactors the ClassroomSettingsModal to use two columns and
reorganizes the most optional settings to the right column.

It also refactors the level editor editable properties list, now
specifying which properties are editable in the schema, differentiated
by product for differences between CodeCombat and Ozaria.

Merges a couple more files together, mostly files that are unused in
Ozaria. Doesn't yet remove the identical copies.

Removes an unused, old classroom `settings` property from an earlier
experimental version of this.