Fix required-but-restricted item logic in InventoryModal
Previously it was possible to get into a state where an item was
required for having one of the properties you needed in the level, but
restricted for not having all of the properties you need in that
slot. Now we avoid requiring an item that doesn't have them all, which
means we can require and point to the first one that does.

Also fix a CoCo/Ozar merge refactoring bug with world map URLs that
was just introduced.