Use another method to initialize/remove fbq to avoid errors (#6895)
Previously, we always loaded and initialized it, but sometimes removed
it while it was initializing, leading to errors when we removed
`window.fbq` mid-load. Now we wait to initialize it until we have
loaded the user and know whether we want to initialize it, and also
wait to remove it at that time.