Add drag-and-drop block support side-by-side with real code (beta)
This uses Google's Blockly to provide a drag-and-drop programming
option, while still keeping the Python, JavaScript, or Lua code
visible, splitting the code editor UI into two halves.

As configured, this is available only to users with beta permissions,
but you can test with `?blocks=opt-out` on any level. Whenever it's
working more reliably, we can change the default from `hidden` to
`opt-in` so that any player can enable it.

For this to move out of work-in-progress state, we need a code ->
Blockly serialization layer to be implemented in Esper. Without that,
all levels start with no code. When that's complete, players will be
able to edit in blocks or in code at the same time, both being kept in
sync.

There's currently an issue with webpack not being able to load the
Blockly locale files, so localization is incomplete. We need to figure
out the right webpack incantation, and then we can uncomment that and
have full i18n support.

Some event handler blocks will be need to be implemented custom,
because Blockly's default function blocks cannot be passes as
values--this would make game development and some pet levels not
work. Blockly is currently disabled for game-dev, web-dev, C++, Java,
and CoffeeScript.

Works pretty well on iPads that don't have physical keyboards.