Add more of Ozaria's syntax highlight coloring, port to CodeCombat
We designed a new colorblind-friendly syntax coloring scheme for
Ozaria, which was fully implemented in chalkboards but only partially
implemented in the code editor, leading to under-colored code. This
adds more of the intended colors (without fully being able to tag
everything we'd want from the chalkboard spec) and makes things more
colorful. It also ports the colors to CodeCombat, which nwo has a
brighter code area background to improve contrast with the code.