Remove Kith & heart seed narrative feature experiment
This reverts #4340, #4339, and #4316, which made a new experimental
feature to add really big LevelDialogueView messages to experiment
with characters narrating level start/success/finish messages in a
core group of dungeon levels.

It was kinda cool, but the UI was always a bit finicky on different
screen sizes, and we hadn't ended up fully testing at the time, having
restricted it to admin only and never gotten around to running the
experiment.

Side effect is that admins saw different stuff than the rest of the
players for a while, and our players were very confused why there was
a random Kithman following them in some levels.

I considered actually turning it fully on and running the experiment,
but we had made many changes to the level progression, and we are
aiming to rework this level progression anyway based on M7, and the UI
layout had only gotten more brittle in the interim.

I've already removed the scripts, heart seeds, and kithmen from all
the dungeon levels, except for leaving one in on Master of Names,
which was kinda cool, even if a bit inexplicable.