Fix occasional 1 FPS level playback bug
I think I finally figured this one out: we had added a separate use of
the idle.js library, which was overwriting some of the callbacks the
main application was using to determine idle/active state. I updated
to a more recent version of it that doesn't clobber its own handlers,
so now we can instantiate both without conflict.

I also made it so that the PlaybackOverScreen doesn't show when we are
paused and playback is not over, for any remaining edge cases that
might reveal arena match winners prematurely.