Stop SegmentedSprites from replaying actions on spritesheet rerender
This is the bug where all the dead sprites would get up and die again
whenever a new animation needed to be rendered that would call for a
new spritesheet. SingularSprites may still do this--didn't figure
that one out--but for the most common case, of SegmentedSprites, they
now pick up where the old animation left off on a rerender, meaning
the dead stay dead!

Satisfying to finally put that one to bed after all these years.