Release level chat for errors generally for home users
Quick and dirty hack to not allow free-form chat (should really clean
up the unneeded code)

Tweaked classroom settings language to be about hints rather than
chat.

Using gpt-4-1106-preview by default now.

Still needing to be done on this:
- Show appropriate messages when quota is up (not subscription modal
for non-home-users)
- Show remaining credits in UI
- Translate error questions into user's locale
- Add a free-form hint button that's not connected to errors
- Don't include previous message conversation history any more now that
it's not a conversation

When we have that stuff, we can just turn it on for student users,
too (since teachers already have the setting in class settings, and it
defaults to 'off').

CCW-2267