Code -> blocks generation (#7210)
Written in collaboration with @basicer, this adds code -> blocks generation for keeping text-based coding and block-based coding in sync. It also adds a basic testing ground for the code -> blocks -> code process at /artisans/block-testing that we can use to make sure all CodeCombat levels can be supported:

This updates Blockly and its plugin versions and introduces patch-package, which was needed for monkey-patching problems in Blockly and its continuous-toolbox plugin.