Refactor session vs. level hero & equipment checks, support Junior
Logic for when we used hero from session vs. level, and when we used
items from those, and when we showed various screens and loaded
various resources, was overly complicated. This refactors it and makes
it support CodeCombat Junior levels.

Because this was so complicated, there could be bugs here.