Move SpellTopBar functionality to GameMenuModal
Since the top bar is hidden now, an alternative way to get to all the
things in there was needed. This adds new tabs to the GameMenuModal
that contain similar functionality to what was previously in the
SpellTopBar.

Once we are satisfied with this, we can get rid of all the SpellTopBar
code.