Fix CCJ-124: auto-snap blocks on most drag events in Junior mode
Doesn't fully snap things together when there start to be block-level
control statements like loops, but before that it should work well.

Future work could include doing this in limited cases in CodeCombat,
in a different way to not always condense newlines in between comments
and intentionally separated lines of code.