Fix CCJ-148: simplify Junior level loading screen and auto-unveil it
Instead of showing goals, loading tip, and intro docs, we just skip
those things and show the level name for CodeCombat Junior. We also
automatically start the level instead of waiting for a Start Level
button press.

This also centers the loading bar and start level button a little
better, for both CodeCombat and CodeCombat Junior.