Fix CCJ-151: clicking customized toolbox block inserts correctly
Now changing a toolbox block's argument (ex.: from "up" to "down")
and then clicking on it will again insert `go("down")` instead of
`go("up")`.

Previously, the customization was lost, because a refactoring broke
the handling that was needed to use it, and a fallback method was used
to make the block that couldn't detected the changed argument.