Fix CCJ-31: allow beta teachers to assign the Junior course
The Ozaria-style teacher dashboard for CodeCombat now allows beta
teachers to assign the CodeCombat Junior course. Beta teachers will
see the Junior course in the curriculum guide, and there will be a
button to add the missing course to the current classroom if they are
looking at a particular classroom.

I think there is currently a bug in the teacher dashboard state update
with that, since it takes a refresh to see any students when looking
at the newly added new Junior course for a particular classroom.

Updated the handling of free and beta courses in various places to
make this work.

Also updated handling of module titles a little bit to allow for
rendering inline <code> blocks for them.

Also updated new teacher dashboard to not use CodeCombat screenshot
thumbnails until we can create new ones, as the old ones had outdated
course timing information on them.

Also updated a few places in new teacher dashboard to specify
difference betwen course (CoCo) and chapter (Oz).