Replace negative lookbehinds in regexes. Fix their bugs.
This brings back some translate-utils tests that got missed in the
decaffeination stage and updates them and their corresponding
translate-utils logic so that they all pass, and then fixes the
regexes to be much simpler and not use negative lookbehinds, which
aren't supported in old versions of Safari.

It also should do a better job of ignoring JS -> Python dictionary key
quoting when it's actually an autocomplete placeholder like ${1:
"foo"}, and same thing for in Lua when those would need to have :
replaced with =. So, we can remove the fixup code in
autocomplete.coffee that was intended to counteract that case. (It
wasn't working in Lua anyway.)

There were also some bugs in translating JavaScript that lacked
semicolons to C++/Java. There probably still are, but now there are
fewer, and the tests pass.