Use a new set of pet heroes in CodeCombat Junior
This hooks up ThangType constants for the new pet heroes, makes the
logic for when to display pet-style hero selection also encompass
whether the campaign is Junior (not just the level), and tries to swap
back to pet heroes when a player would get into a CodeCombat Junior
level with a normal warrior hero.

There's also a start of handling student hero choice between heroes in
the student dashboard, but that'll need more work once we really know
whether a student is just going to be doing CCJ or also other courses
in advance, on account creation.