Add CodeCombat Junior home beta experiment
When new, anonymous users meeting certain criteria land on /play, they
have a chance to see a simple modal showing the choice of either
playing CodeCombat Junior or CodeCombat.

This modal's content is a bit of a stub for now, although it can work
better than nothing.

Also refactors a few other things in CampaignView and related
files, like some of the extra decaffeinate returns and null checks.

Also hides the anonymous classroom modal in more cases, like small
screens, when you've closed it before, and when you are on the Junior
campaign.