Fix CampaignEditor Treema nodes for Campaigns (#7457)
We had two implementations for CampaignsNodes for setting adjacent
campaigns, but neither worked because of merging typos. Fixed typos
and condensed down to one.

Tried doing similar for LevelsNodes, but turned out we needed
different approaches (assuming Ozaria approach works for Ozaria; it
didn't for CodeCombat). Refactored a bit there.