Managed CCJ paid levels via level.requiresSubscription
Instead of hardcoding a list of free levels, we now just go back to
the original way of doing it, where paid levels have
`level.requiredSubscription` set to `true`. This is also changed on
the server.

This updates the CampaignView to handle this for both home version and
classroom version. It also fixes some issues displaying practice level
play buttons, especially for home version.

Separately, on /play/level/<slug>, loading errors now just show that
you need a license or subscription, not also showing that there was an
error loading the level.

Home version was showing level-specific score leaderboards in Junior,
which we don't want (too complicated, apparently not really working
anyway).