Update victory modals to handle Junior practice levels
Home version modal does not show practice levels as next levels if the
home user is not paid, since those are all paid.

Classroom version does similarly. Also, when it shows a paid level
next and there is no license, it will navigate to that level's page so
that the license requirement message can be shown. (This was
previously not handled for partially-paid courses and classroom items
& gems mode.)

Fixed a bug trying to show nonexistent Junior hero images in the
classroom victory modal.