Fix InventoryModal's auto-unlocks of required items (#7696)
Previously, we didn't auto-unlock items that were paid. Also, we
weren't the best at picking _which_ item to auto-unlock, sometimes
leading to an incorrect recommendation.

Now we unlock the earliest free item that can match all the
requirements; failing that, the earliest paid item. If we have an
item that meets the requirements, we will point to the best owned
item. We'll ignore kinda unreleased items.

Also refactor a few misc decaffeinated Array.from calls.