Automatically equip required items when in a level without them
If #7696 helps solve an issue where we don't own the proper items when
going into a level in the InventoryModal, this solves similar issues
if we somehow find ourselves in a level without the proper items (like
navigating there by direct level link, or some other buggy flow).

This checks for requiredProperties, requiredGear, and restrictedGear,
and does some just-in-time adjustments to what we have equipped in our
session's heroConfig.inventory to meet the requirements.

There are probably some edge cases that aren't handled perfectly, like
unequipping stuff in restricted slots that the default level config
doesn't equip anything for, or dealing with rings, but this should do
a good-enough job at the common cases for a fallback defense
system. More edge cases are handled in the InventoryModal logic, but
that's overkill complexity for here.