Handle free/paid levels for CodeCombat Junior (#7691)
* Managed CCJ paid levels via level.requiresSubscription

Instead of hardcoding a list of free levels, we now just go back to
the original way of doing it, where paid levels have
`level.requiredSubscription` set to `true`. This is also changed on
the server.

This updates the CampaignView to handle this for both home version and
classroom version. It also fixes some issues displaying practice level
play buttons, especially for home version.

Separately, on /play/level/<slug>, loading errors now just show that
you need a license or subscription, not also showing that there was an
error loading the level.

Home version was showing level-specific score leaderboards in Junior,
which we don't want (too complicated, apparently not really working
anyway).

* Update victory modals to handle Junior practice levels

Home version modal does not show practice levels as next levels if the
home user is not paid, since those are all paid.

Classroom version does similarly. Also, when it shows a paid level
next and there is no license, it will navigate to that level's page so
that the license requirement message can be shown. (This was
previously not handled for partially-paid courses and classroom items
& gems mode.)

Fixed a bug trying to show nonexistent Junior hero images in the
classroom victory modal.

* Minor styling and content fixes for SubscribeModal

Also fixed a couple typos