Fix CCJ-269: fix Blockly garbling Unicode comments with quotes
Blockly would start to mess up its string encoding if it saw a field
value with non-ASCII text that had double quotes in them. Example was
CS1 L14 Cupboards of Kithgard in Chinese or French. To work around, we
rewrite the double quotes in comments as single quotes if the comments
have non-ASCII text, before Blockly sees it.