Fix CCJ-123: add "close" button to victory modal
Also makes it so that keyboard focus goes to "continue" button in
victory modal, which lets you press escape to close the modal, instead
of staying in the ace editor.

Fixes a random extra box shadow in CourseVictoryModal that was causing
minor visual glitching.

Trivial refactor to some logging code that was decaffeinated poorly
and logging in two cases instead of one.