Fix CCJ-43: add TTS pronouncing blocks on insert
When you add a new block to the workspace, it uses TTS to pronounce
the block, for CodeCombat Junior.

Probably there are some blocks which aren't pronounced well, and need
to see how this interactions with the block i18n thing to get that
working.

It would be good to preload this a bit so that they happen faster on
first insertion.

TBD on whether players like this, but could be good for low-readers,
especially with directional arrow confusions.