Fix CCJ-110: delete blocks by dragging off screen
By default, Blockly just deletes blocks that you drag back to the
workspace on the right side, or to the trashcan (which we are not
using). If you drag off the other three sides, the blocks just kinda
live out there, invisible, in the void, until something snaps the
blocks back to the center. Confusing!

Now, we check if the dragged block would be (close to) off the screen,
so you can drag off the other three sides and delete the block.