Lower interpreter statement execution limit for CCJ levels
CodeCombat levels interpret a max of 3M statements before throwing an
error, which can take a while to trigger on infinite loops withs
slower hardware.

CodeCombat Junior levels should be much simpler as a rule, and are
fairly prone to infinite loops that don't have automatic yields
inserted in them once we get into the later modules that use while
loops, so this sets the limit to 100K statements, which should trigger
on an infinite loop quite quickly.