Hack to support dynamic `health` global in CCJ levels
Instead of assigning `health` -> `hero.health` statically at the start
of the program, like is done in the programming.Programmable
Component, we instead rewrite any bare instances of `health` to
`hero.health` in a way that's hopefully invisible to the player, so
that references to `health` will get the current value.

Otherwise, code like `while health < 4:` does not work, because
`health` is initialized to `4` and never updated as the hero takes
damage.

Couldn't find another way around this given all the voodoo we do
around making these globals work for CCJ, but seems to do fine in my
testing, and should be strictly scoped to CCJ and the `health`
property.