Fix CCJ-288: fix bugs with heart vs. chicken goal icons
I had started with a `saveThangs` length condition and kept making it
more specific, but actually the `saveThangs` condition was problematic
in the first place, and the later conditions were sufficient.

This also makes it such that we don't show a "current / total" suffix
on the health goal using a similar condition.