Make generated Junior blocks always connect to each other
Previously, we would sometimes put them in separate stacks, and if
they were inside another block, that would also involve adding a
newline block.

I don't remember exactly why we did this, but I think it had something
to do with CodeCombat and newlines and comments and such things being
important.

In any case, we don't seem to need it in CodeCombat Junior, and this
code change seems to do the trick, while not having any effect outside
of Junior.

Able to remove a long function that was counteracting these newline
blocks for maxBlocks counting in CCJ as well. (That I just added
today.)

Fixes CCJ-285. Fixes CCJ-293.