Fix CCJ-306: make CCJ hero fall into water when going off path
Instead of moving into a wall, saying "I can't get there", and then
maybe being off path, we just drown the hero.

The code for this is all in the game engine, but noting it here as we
are adding a new Splash requiredThangType to the JuniorPlayer config
to show the VFX for it.