Fix CCJ-298: adjust code language/format selections on mobile
Recently-ish, we made it so that you could only pick blocks-text and
blocks-icons on iPad. But it didn't work because the iPad detection
was wrong.

That didn't really make sense, because we only support text-code for
CodeCombat, and that needed to apply only to CCJ.

But also, some people have physical keyboards on their iPads, and
since we can't easily detect that, we should still let them choose
text code options.

So, I made it correctly detect mobile, but just for CCJ, and exempting
iPads. Probably it should exempt all tablets or devices that might
have physical keyboards. For example, is this going to turn off
keyboard options in CCJ for Chromebooks or PCs with touch screens?

The other change here is to use the updated mobile detection to
determine when to skip over the `fancySelect` changing the language
and code format dropdowns on mobile, since the other way was too old
and not doing it, leading to duplicated selects (one of which didn't
work).