Add placeholder comparator/var block icon images
These aren't actually used because I don't have a good way to put
them in, but in case we do want to use them at some point, they can be
a starting point. We do have similar versions of these as pickup icons
uploaded as Thangs to be obtained in precursor levels to these
concepts being introduced. If we really want to use these more, we
probably want to get better versions done.